botIATelegram-01/requirements.txt
pypdf
langchain
langchain-community
langchain-google-genai
langchain-chroma
python-telegram-bot
python-dotenv
flask
langchain-groq
langchain-classic
